+++ vdr/api/v1/api_test.go
@@ -306,8 +306,7 @@ func TestWrapper_UpdateDID(t *testing.T) {
 		ID: *id,
 	}
 	didUpdate := DIDUpdateRequest{
-		Document:    *didDoc,
-		CurrentHash: "452d9e89d5bd5d9225fb6daecd579e7388a166c7661ca04e47fd3cd8446e4620",
+		Document: *didDoc,
 	}
 
 	t.Run("ok", func(t *testing.T) {