+++ vdr/api/v1/client_test.go
@@ -140,12 +140,11 @@ func TestHTTPClient_Update(t *testing.T) {
 	didDoc := did.Document{
 		ID: vdr.TestDIDA,
 	}
-	hash := "0000000000000000000000000000000000000000"
 
 	t.Run("ok", func(t *testing.T) {
 		s := httptest.NewServer(&http2.Handler{StatusCode: http.StatusOK, ResponseData: didDoc})
 		c := getClient(s.URL)
-		doc, err := c.Update(vdr.TestDIDA.String(), hash, didDoc)
+		doc, err := c.Update(vdr.TestDIDA.String(), didDoc)
 		require.NoError(t, err)
 		assert.NotNil(t, doc)
 	})
@@ -154,7 +153,7 @@ func TestHTTPClient_Update(t *testing.T) {
 		s := httptest.NewServer(&http2.Handler{StatusCode: http.StatusNotFound, ResponseData: ""})
 		c := getClient(s.URL)
 
-		_, err := c.Update(vdr.TestDIDA.String(), hash, didDoc)
+		_, err := c.Update(vdr.TestDIDA.String(), didDoc)
 
 		assert.Error(t, err)
 	})
@@ -163,14 +162,14 @@ func TestHTTPClient_Update(t *testing.T) {
 		s := httptest.NewServer(&http2.Handler{StatusCode: http.StatusOK, ResponseData: "}"})
 		c := getClient(s.URL)
 
-		_, err := c.Update(vdr.TestDIDA.String(), hash, didDoc)
+		_, err := c.Update(vdr.TestDIDA.String(), didDoc)
 
 		assert.Error(t, err)
 	})
 
 	t.Run("error - wrong address", func(t *testing.T) {
 		c := getClient("not_an_address")
-		_, err := c.Update(vdr.TestDIDA.String(), hash, didDoc)
+		_, err := c.Update(vdr.TestDIDA.String(), didDoc)
 		assert.Error(t, err)
 	})
 }